Switch to int64_t throughout most of the codebase; This resolves some errors resulting from the use of unsigned types--these were exposed by Visual Studio, but hidden by Clang/GCC; Suppress tons of errors;

+++ src/carli/action.h
@@ -39,7 +39,7 @@ namespace Carli {
       return oss.str();
     }
 
-    virtual int compare(const Action &rhs) const = 0;
+    virtual int64_t compare(const Action &rhs) const = 0;
 
     virtual void print_impl(std::ostream &os) const = 0;
   };
